backend/requirements.txt
sqlmodel
asyncpg
python-dotenv
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
